+++ CHANGELOG.md
@@ -17,6 +17,8 @@
 
 ### Upcoming Release
 
+* [#251] [FEATURE] Raise a helpful error when an attribute is missing from
+  `ATTRIBUTE_TYPES`
 * [#247] [BUGFIX] Populate `has_many` and `belongs_to` select boxes
   with the current value of the relationship.
 * [#237] [I18n] Fix broken paths for several I18n files (de, es, fr, pt-BR, vi).